requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
pytz>=2024.1
PyYAML>=6.0.2
